+++ HotelWise.Domain/Validator/HotelValidator.cs
@@ -9,7 +9,7 @@ namespace HotelWise.Domain.Validator
         {
             RuleFor(h => h.HotelName).NotEmpty().MaximumLength(100);
             RuleFor(h => h.Description).MaximumLength(500);
-            RuleFor(h => h.Stars).InclusiveBetween(1, 5);
+            RuleFor(h => (int)h.Stars).InclusiveBetween(1, 5);
             RuleFor(h => h.InitialRoomPrice).GreaterThan(0);
             RuleFor(h => h.ZipCode).MaximumLength(10);
             RuleFor(h => h.Location).MaximumLength(200);
